+draft of eva predictions
Ignore-this: 90577f5f7b317a04e612c29fdb32e090
This is a compilation of predictions related to the new Eva movies.
The idea is to calibrate myself and see what theories are full of it and what not.
This may tie in to otaku-essay.page or may not; we'll see by no later than 2015 (probably).

darcs-hash:20110303220011-f7719-f7690c2a9a1f6f707b14d229325b35bcb084586b