finished _The Thousand Autumns of Jacob de Zoet_ tonight
Ignore-this: 3b967c34080597dfea55c81cc951cc28
Finally got around to reading it. It was surprisingly unliterary
and unpostmodern for Mitchell, but in exchange, he nailed the historical
details and gave us an adventure which subverted many of the usual
tropes - the raid on the nunnery was just a trap, the hero doesn't get
the girl, his chief heroism was standing there to be shot at, and the
man who takes down the big baddie is someone we thought to be entirely in the
baddie's pocket. The supernatural aspects are implied to be genuine, but it's
never resolved, which I am grateful for.

rolling back:

Fri Jan  6 00:08:02 EST 2012  gwern0@gmail.com
  * Books: +mitchell's latest

    M ./Book reading list.page +1

darcs-hash:20120106051108-af090-fb1c6b1b01cd85f22ce8201dc5c30cf3de003e73